# Changes: - cart

+++ frontend/src/app/components/dish-list/dish-list.component.ts
@@ -2,6 +2,7 @@ import {Component, OnInit} from '@angular/core';
 import {DishService} from "../../services/dish/dish.service";
 import {faPlus, faMinus, faDollar, faTrashCan} from "@fortawesome/free-solid-svg-icons";
 import {Dish} from "../../models/Dish";
+import {CartService} from "../../services/cart/cart.service";
 
 
 @Component({
@@ -30,11 +31,11 @@ export class DishListComponent implements OnInit {
 
   searchRatings: number[] = [];
 
-  searchMinPrice: number[]=[];
+  searchMinPrice: number[] = [];
 
-  searchMaxPrice: number[]=[];
+  searchMaxPrice: number[] = [];
 
-  constructor(private dishService: DishService) {
+  constructor(private dishService: DishService, private cartService: CartService) {
   }
 
   ngOnInit() {
@@ -43,10 +44,12 @@ export class DishListComponent implements OnInit {
 
   getDishes(): void {
     this.dishService.findAll().subscribe(dishes => {
-      console.log(dishes);
         this.dishes = dishes;
         dishes.forEach(dish => {
-          dish.available = dish.quantity;
+          this.cartService.getPortionOfDishInCarts(1, dish.dishId).subscribe(portion => {
+            console.log(portion);
+            dish.reserved = portion;
+          });
           if (dish.price > this.mostExpensiveValue) {
             this.mostExpensiveValue = dish.price;
           }
@@ -58,21 +61,26 @@ export class DishListComponent implements OnInit {
     )
   }
 
-  public addItem(dish: any) {
-    if (dish.available > 0) {
-      dish.available--;
+  public addItem(dish: Dish) {
+    if (dish.quantity > 0) {
+      dish.quantity--;
+      dish.reserved++;
+      this.cartService.addToCart(1, dish.dishId, 1).subscribe();
     }
   }
 
   public removeItem(dish: any) {
-    if (dish.available < dish.amount) {
-      dish.available++;
+    if (dish.reserved > 0) {
+      dish.reserved--;
+      dish.quantity++;
+      this.cartService.removeFromCart(1, dish.dishId, 1).subscribe();
     }
   }
 
   public delete(dish: Dish) {
-    this.dishes = this.dishes.filter(d => d !== dish);
-    this.dishService.deleteDish(dish.dishId).subscribe();
+    if (confirm("Are you sure to delete " + dish.name)) {
+      this.dishes = this.dishes.filter(d => d !== dish);
+      this.dishService.deleteDish(dish.dishId).subscribe();
+    }
   }
-
 }
